chore: api 더미데이터 파일 이동 (#12)
