+++ src/constants/dummy.tsx
@@ -1,0 +1,145 @@
+export const INTERVIEW_QUESTION = [
+    { interview_question_id: 1, title: '프로세스의 구조에 대해 말씀해주세요' },
+    {
+        interview_question_id: 2,
+        title: '프로세스와 스레드의 차이에 대해 말씀해주세요',
+    },
+    {
+        interview_question_id: 3,
+        title: '네트워크 OSI 7계층에 대해서 말씀해주세요',
+    },
+    { interview_question_id: 4, title: 'TCP와 UDP의 차이에 대해 말씀해주세요' },
+    {
+        interview_question_id: 5,
+        title: 'TCP의 3way handshake에 대해 말씀해주세요',
+    },
+    { interview_question_id: 6, title: 'REST API에 대해 설명해주세요' },
+    {
+        interview_question_id: 7,
+        title: '클라이언트/서버 모델에 대해 말씀해주세요',
+    },
+    { interview_question_id: 8, title: 'CSRF 토큰에 대해 말씀해주세요' },
+    {
+        interview_question_id: 9,
+        title: 'Same Orogin Policy에 대해 말씀해주세요',
+    },
+    {
+        interview_question_id: 10,
+        title: '브라우저 렌더링 원리에 대해 말씀해주세요',
+    },
+];
+
+export const INTERVIEW_RESULT = [
+    {
+        question_id: 1, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요 1 ',
+        additional_question_2:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요 2 ',
+        additional_question_3:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요 3',
+    },
+    {
+        question_id: 2, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요1',
+        additional_question_2:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요2',
+        additional_question_3:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요3',
+    },
+    {
+        question_id: 3, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요1',
+        additional_question_2:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요2',
+        additional_question_3:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요3',
+    },
+    {
+        question_id: 4, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요1',
+        additional_question_2:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요2',
+        additional_question_3:
+            '답변 감사합니다. OSI 7계층 중 전송계층에 대해서 자세하게 설명해주세요3',
+    },
+    {
+        question_id: 5, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1: 'additional_question11',
+        additional_question_2: 'additional_question22',
+        additional_question_3: 'additional_question33',
+    },
+    {
+        question_id: 6, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1: 'additional_question1',
+        additional_question_2: 'additional_question2',
+        additional_question_3: 'additional_question3',
+    },
+    {
+        question_id: 7, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1: 'additional_question1',
+        additional_question_2: 'additional_question2',
+        additional_question_3: 'additional_question3',
+    },
+    {
+        question_id: 8, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1: 'additional_question1',
+        additional_question_2: 'additional_question2',
+        additional_question_3: 'additional_question3',
+    },
+    {
+        question_id: 9, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1: 'additional_question1',
+        additional_question_2: 'additional_question2',
+        additional_question_3: 'additional_question3',
+    },
+    {
+        question_id: 10, // 고유 번호 1~10
+        question: '네트워크 OSI 7계층에 대해서 설명해주세요',
+        answer: '프로세스의 구조는 data, stack, code로 구성됐습니다.', //내 답변
+        gpt_answer:
+            '프로세스의 구조는 data, heap, stack, code 총 4가지로 구성됐습니다.',
+        additional_question_1: 'additional_question1',
+        additional_question_2: 'additional_question2',
+        additional_question_3: 'additional_question3',
+    },
+];
